LPD-79147 Adapt permission management

+++ modules/apps/commerce/commerce-inventory-service/src/main/java/com/liferay/commerce/inventory/service/impl/CommerceInventoryWarehouseItemServiceImpl.java
@@ -11,7 +11,6 @@ import com.liferay.commerce.inventory.model.CommerceInventoryWarehouseItem;
 import com.liferay.commerce.inventory.service.base.CommerceInventoryWarehouseItemServiceBaseImpl;
 import com.liferay.portal.aop.AopService;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
 import com.liferay.portal.kernel.security.permission.resource.PortletResourcePermission;
@@ -104,17 +103,22 @@ public class CommerceInventoryWarehouseItemServiceImpl
 			long companyId, String sku, String unitOfMeasureKey)
 		throws PortalException {
 
-		PortletResourcePermission portletResourcePermission =
-			_commerceInventoryWarehouseModelResourcePermission.
-				getPortletResourcePermission();
+		for (Long commerceInventoryWarehouseId :
+				commerceInventoryWarehouseItemLocalService.
+					getCommerceInventoryWarehouseIds(
+						companyId, sku, unitOfMeasureKey)) {
 
-		portletResourcePermission.check(
-			getPermissionChecker(), null,
-			CommerceInventoryActionKeys.MANAGE_INVENTORY);
+			if (!_commerceInventoryWarehouseModelResourcePermission.contains(
+					getPermissionChecker(), commerceInventoryWarehouseId,
+					ActionKeys.UPDATE)) {
 
-		commerceInventoryWarehouseItemLocalService.
-			deleteCommerceInventoryWarehouseItems(
-				companyId, sku, unitOfMeasureKey);
+				continue;
+			}
+
+			commerceInventoryWarehouseItemLocalService.
+				deleteCommerceInventoryWarehouseItems(
+					commerceInventoryWarehouseId);
+		}
 	}
 
 	@Override
@@ -214,7 +218,7 @@ public class CommerceInventoryWarehouseItemServiceImpl
 
 		portletResourcePermission.check(
 			getPermissionChecker(), null,
-			CommerceInventoryActionKeys.MANAGE_INVENTORY);
+			CommerceInventoryActionKeys.VIEW_INVENTORIES);
 
 		return commerceInventoryWarehouseItemLocalService.
 			getCommerceInventoryWarehouseItemsByCompanyId(
@@ -223,23 +227,13 @@ public class CommerceInventoryWarehouseItemServiceImpl
 
 	@Override
 	public List<CommerceInventoryWarehouseItem>
-			getCommerceInventoryWarehouseItemsByCompanyIdSkuAndUnitOfMeasureKey(
-				long companyId, String sku, String unitOfMeasureKey, int start,
-				int end)
-		throws PrincipalException {
-
-		PortletResourcePermission portletResourcePermission =
-			_commerceInventoryWarehouseModelResourcePermission.
-				getPortletResourcePermission();
-
-		boolean replacePermissionCheck = !portletResourcePermission.contains(
-			getPermissionChecker(), null,
-			CommerceInventoryActionKeys.MANAGE_INVENTORY);
+		getCommerceInventoryWarehouseItemsByCompanyIdSkuAndUnitOfMeasureKey(
+			long companyId, String sku, String unitOfMeasureKey, int start,
+			int end) {
 
 		return commerceInventoryWarehouseItemLocalService.
 			getCommerceInventoryWarehouseItemsByCompanyIdSkuAndUnitOfMeasureKey(
-				companyId, sku, unitOfMeasureKey, start, end,
-				replacePermissionCheck);
+				companyId, sku, unitOfMeasureKey, start, end, true);
 	}
 
 	@Override
@@ -268,7 +262,7 @@ public class CommerceInventoryWarehouseItemServiceImpl
 
 		portletResourcePermission.check(
 			getPermissionChecker(), null,
-			CommerceInventoryActionKeys.MANAGE_INVENTORY);
+			CommerceInventoryActionKeys.VIEW_INVENTORIES);
 
 		return commerceInventoryWarehouseItemLocalService.
 			getCommerceInventoryWarehouseItemsCount(
@@ -280,16 +274,9 @@ public class CommerceInventoryWarehouseItemServiceImpl
 			long companyId, String sku, String unitOfMeasureKey)
 		throws PortalException {
 
-		PortletResourcePermission portletResourcePermission =
-			_commerceInventoryWarehouseModelResourcePermission.
-				getPortletResourcePermission();
-
 		return commerceInventoryWarehouseItemLocalService.
 			getCommerceInventoryWarehouseItemsCount(
-				companyId, sku, unitOfMeasureKey,
-				!portletResourcePermission.contains(
-					getPermissionChecker(), null,
-					CommerceInventoryActionKeys.MANAGE_INVENTORY));
+				companyId, sku, unitOfMeasureKey, true);
 	}
 
 	@Override
@@ -303,7 +290,7 @@ public class CommerceInventoryWarehouseItemServiceImpl
 
 		portletResourcePermission.check(
 			getPermissionChecker(), null,
-			CommerceInventoryActionKeys.MANAGE_INVENTORY);
+			CommerceInventoryActionKeys.VIEW_INVENTORIES);
 
 		return commerceInventoryWarehouseItemLocalService.
 			getCommerceInventoryWarehouseItemsCountByCompanyId(companyId);
@@ -312,7 +299,7 @@ public class CommerceInventoryWarehouseItemServiceImpl
 	@Override
 	public int getCommerceInventoryWarehouseItemsCountByModifiedDate(
 			long companyId, Date startDate, Date endDate)
-		throws PrincipalException {
+		throws PortalException {
 
 		PortletResourcePermission portletResourcePermission =
 			_commerceInventoryWarehouseModelResourcePermission.
@@ -320,7 +307,7 @@ public class CommerceInventoryWarehouseItemServiceImpl
 
 		portletResourcePermission.check(
 			getPermissionChecker(), null,
-			CommerceInventoryActionKeys.MANAGE_INVENTORY);
+			CommerceInventoryActionKeys.VIEW_INVENTORIES);
 
 		return commerceInventoryWarehouseItemLocalService.
 			getCommerceInventoryWarehouseItemsCountByModifiedDate(
@@ -332,7 +319,7 @@ public class CommerceInventoryWarehouseItemServiceImpl
 			getCommerceInventoryWarehouseItemsCountByModifiedDate(
 				long companyId, Date startDate, Date endDate, int start,
 				int end)
-		throws PrincipalException {
+		throws PortalException {
 
 		PortletResourcePermission portletResourcePermission =
 			_commerceInventoryWarehouseModelResourcePermission.
@@ -340,7 +327,7 @@ public class CommerceInventoryWarehouseItemServiceImpl
 
 		portletResourcePermission.check(
 			getPermissionChecker(), null,
-			CommerceInventoryActionKeys.MANAGE_INVENTORY);
+			CommerceInventoryActionKeys.VIEW_INVENTORIES);
 
 		return commerceInventoryWarehouseItemLocalService.
 			getCommerceInventoryWarehouseItemsByModifiedDate(
